Nu main program, replaced AppController with Nu class.
main.m, AppController.[hm], and Scriptability.h are now empty.
The contents of Scriptability.h were moved to the .m file
since this header wasn't used anywhere else, and eventually
Scriptability.m will also collapse into Nu.

main.nu now starts the application and contains a definition
of AppController.  As a bonus, it also adds a menu item
to enable the interactive Nu console, so that we can
interact with our classes and objects at runtime.